digitalbook/requirements.txt
Django>=4.2.0,<5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
# corsheaders (invalid package name) removed; using django-cors-headers below
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.6,<3.0.0
dj-database-url>=1.2.0,<2.0.0
drf-nested-routers>=0.93.4,<1.0.0
Pillow>=10.0.0
whitenoise>=6.0.0
gunicorn>=20.1.0
django-cors-headers>=4.0.0,<5.0.0
# Dépendances pour le parsing PDF
PyMuPDF>=1.22.0
pdfplumber>=0.9.0

# OCR pour les pages scannées (utilisé par scripts/algo_balise.py)
pytesseract>=0.3.10

# Dépendances pour la génération de QCM avec IA
openai>=1.0.0

# Tâches asynchrones
celery>=5.3,<6.0
redis>=5.0.0
